fix: code coverage
