+++ packages/shopware-6-client/__tests__/helpers/convertAssociations.spec.ts
@@ -1,0 +1,8 @@
+import { convertAssociations } from "../../src/helpers/convertAssociations";
+
+describe("SearchConverter - convertAssociations", () => {
+  it("should returns empty object if no params provided", () => {
+    const result = convertAssociations();
+    expect(result).toBeUndefined();
+  });
+});
